backend/requirements.txt
langgraph
langchain
langchain-openai
langchain-community
tavily-python
streamlit
python-dotenv
pydantic
langchain-google-genai
